ARROW-5747: [C++] Improve header and column names options
The `header_rows` option is removed (it seemed untested and probably didn't work well)
and replaced with a `skip_rows` option that dictates how many rows are skipped at
the start of the CSV file (default 0).

Furthermore, a `column_names` option is added to hardcode the column names.
If empty (default), they are read from the first non-skipped row in the CSV file.